requirements.txt
@"
fastapi==0.110.0
uvicorn==0.29.0
transformers==4.40.0
torch==2.2.2
python-spellchecker
spacy==3.7.4
pydantic>=2.0
python-multipart==0.0.9
"@ | Out-File -Encoding utf8 requirements.txt